Revert "WIP: move some things to url-utils."
This reverts commit ff54f559013186e001362ccbbc9b44bb0a33bd0a.
